Change default value of DYNAMIC_ARCH to True
Since this is the version we want compiled in bintray, it is better to be the
default value in order to not require all libraries that depend on openblas to
specify this option.

If the user is compiling the recipe in his machine he can then pass the value
False to DYNAMIC_ARCH to avoid compiling for many different architectures in his
machine.